requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pytesseract
transformers
protobuf
Pillow
python-dotenv